Move `ChannelInputModel` to `channels.py`
