+++ tests/tasks/test_registration.py
@@ -14,7 +14,7 @@ from PIL import Image
 from pytest import MonkeyPatch
 
 from fractal_tasks_core.ngff.zarr_utils import load_NgffImageMeta
-from fractal_tasks_core.ome_zarr.input_models import ChannelInputModel
+from fractal_tasks_core.ome_zarr.channels import ChannelInputModel
 from fractal_tasks_core.roi import (
     convert_indices_to_regions,
 )
